First commit
Created a small size python program to keep track of stocks bought and sold. Used csv and pandas module to take care of the csv file and the data-frame. This program doesn't follow the OOP paradigm. Maybe sometime in the future, I will rewrite the entire program by following the OOP paradigm.